Tweak behavior of redirect for http codes
Not every webserver supports 308 redirects with POST correctly.
Most sites are already configured using 301 or 302 redirects.
This only effects POST, GET behavior is unmodified.
Most calls to POST are under control by friendica not the end user,
so should not be a problem ni general.

GET with 301, 302, 303, 307 and 308 are redirected as GET
POST with 301, 302, 307 and 308 are redirected as POST
POST with 303 is redirected as GET