Hope I have repaired the channel admin page.
Channel blocking and deleting was copied from user actions. This was not
done to an end. Hope what I do is enough to enable channel blocking and
deleting the correct way.

On deletion all entities that belong to the channel are deleted. But the
channel itself is just marked as deleted. Do not really understand why
it is done this way.