Extended theme redbasic: Allow to center the content (aside + converse)
Added an option to settings of redbasic theme. You are now able to
activate centering the left aside and the conversion regions in the
middle of the screen. This respects configuration of conversion region
width and works without special actions for mobile devices.

TODO: Find a way to handle conversion width given in percent.