Bootstrapify item tools
Only show the like buttons and the delete button permanently hide
everything else in a drop down. This will break non bootstrap themes.
So may be we also want to move this into the red basic folder. It still
a little bit work in progress. Allignment is not 100% correct. May be
we want to add text into the drop down.