Scripts to add external theme and addon repositories. Example:     util/add_theme_repo https://github.com/deadsuperhero/redmatrix-themes deadsuperhero
The final argument is a nickname on this system. The repository will be placed under extend/theme/deadsuperhero and the themes from that repository linked to your redmatrix theme directory.

You can do the same with addons. We'll have to come up with some "pull" scripts to keep updated but the clever folks will already know how to do this.