View upload photo button only if you have permission.
Check if post_photos permission is available otherwise hide the photo
upload button.
Or should we show the button and just disable it with a title, about
missing permission to upload photos?
In comments we use the same icon, but just add [img] bb-code. Should we
provide this if no permission to upload photos. Do we need different
icons for [img] and photo upload buttons?